+++ lib/benchling-webhook-stack.ts
@@ -1,10 +1,9 @@
 import * as cdk from "aws-cdk-lib";
-import * as s3 from "aws-cdk-lib/aws-s3";
 import * as ec2 from "aws-cdk-lib/aws-ec2";
 import * as ecr from "aws-cdk-lib/aws-ecr";
 import { Construct } from "constructs";
 import { FargateService } from "./fargate-service";
-import { AlbApiGateway } from "./alb-api-gateway";
+import { HttpApiGateway } from "./http-api-gateway";
 import { ProfileConfig } from "./types/config";
 import packageJson from "../package.json";
 
@@ -23,9 +22,8 @@ export interface BenchlingWebhookStackProps extends cdk.StackProps {
 }
 
 export class BenchlingWebhookStack extends cdk.Stack {
-    private readonly bucket: s3.IBucket;
     private readonly fargateService: FargateService;
-    private readonly api: AlbApiGateway;
+    private readonly api: HttpApiGateway;
     public readonly webhookEndpoint: string;
 
     constructor(
@@ -148,8 +146,6 @@ export class BenchlingWebhookStack extends cdk.Stack {
 
         // Bucket name will be resolved at runtime from CloudFormation outputs
         // For CDK purposes, we use a placeholder for IAM permissions
-        this.bucket = s3.Bucket.fromBucketName(this, "BWBucket", "placeholder-bucket-resolved-at-runtime");
-
         // Get the default VPC or create a new one
         const vpc = ec2.Vpc.fromLookup(this, "DefaultVPC", {
             isDefault: true,
@@ -172,7 +168,6 @@ export class BenchlingWebhookStack extends cdk.Stack {
         // Build Fargate Service props using new config structure
         this.fargateService = new FargateService(this, "FargateService", {
             vpc,
-            bucket: this.bucket,
             config: config,
             ecrRepository: ecrRepo,
             imageTag: imageTagValue,
@@ -196,14 +191,16 @@ export class BenchlingWebhookStack extends cdk.Stack {
             logLevel: logLevelValue,
         });
 
-        // Create API Gateway that routes to the ALB
-        this.api = new AlbApiGateway(this, "ApiGateway", {
-            loadBalancer: this.fargateService.loadBalancer,
+        // Create HTTP API that routes through VPC Link to the service
+        this.api = new HttpApiGateway(this, "HttpApiGateway", {
+            vpc: vpc,
+            cloudMapService: this.fargateService.cloudMapService,
+            serviceSecurityGroup: this.fargateService.service.connections.securityGroups[0],
             config: config,
         });
 
         // Store webhook endpoint for easy access
-        this.webhookEndpoint = this.api.api.url;
+        this.webhookEndpoint = this.api.api.url!;
 
         // Export webhook endpoint as a stack output
         new cdk.CfnOutput(this, "WebhookEndpoint", {